requirements.txt
discord.py>=2.3.2
flask>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
feedparser>=6.0.11
Pillow>=10.0.0
